Refactor app structure
