+++ retrorecon/routes/db.py
@@ -10,7 +10,7 @@ def new_db():
     safe = app._sanitize_db_name(name)
     if not safe:
         flash('Invalid database name.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     app.close_connection(None)
     temp_path = os.path.join(app.get_db_folder(), app.TEMP_DB_NAME)
     if app.app.config.get('DATABASE') == temp_path and os.path.exists(temp_path):
@@ -21,7 +21,7 @@ def new_db():
         flash('New database created.', 'success')
     except ValueError as e:
         flash(str(e), 'error')
-    return redirect(url_for('index'))
+    return redirect(url_for('core.index'))
 
 
 @bp.route('/load_db', methods=['POST'])
@@ -29,11 +29,11 @@ def load_db_route():
     file = request.files.get('db_file')
     if not file:
         flash("No database file uploaded.", "error")
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     filename = app._sanitize_db_name(file.filename or '')
     if not filename:
         flash('Invalid database file.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     db_path = os.path.join(app.get_db_folder(), filename)
     app.close_connection(None)
     temp_path = os.path.join(app.get_db_folder(), app.TEMP_DB_NAME)
@@ -47,14 +47,14 @@ def load_db_route():
         flash("Database loaded.", "success")
     except Exception as e:
         flash(f"Error loading database: {e}", "error")
-    return redirect(url_for('index'))
+    return redirect(url_for('core.index'))
 
 
 @bp.route('/save_db', methods=['GET'])
 def save_db():
     if not app._db_loaded():
         flash('No database loaded.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     name = request.args.get("name", "").strip()
     if name:
         safe_name = app._sanitize_export_name(name)
@@ -73,22 +73,22 @@ def rename_db():
     safe = app._sanitize_db_name(new_name or '')
     if not safe:
         flash('Invalid database name.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     if not app._db_loaded():
         flash('No database loaded.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     app.close_connection(None)
     new_path = os.path.join(app.get_db_folder(), safe)
     try:
         os.rename(app.app.config['DATABASE'], new_path)
     except OSError as e:
         flash(f'Error renaming database: {e}', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     app.app.config['DATABASE'] = new_path
     app.ensure_schema()
     session['db_display_name'] = safe
     flash('Database renamed.', 'success')
-    return redirect(url_for('index'))
+    return redirect(url_for('core.index'))
 
 
 @bp.route('/load_saved_db', methods=['POST'])
@@ -98,11 +98,11 @@ def load_saved_db():
     safe = app._sanitize_db_name(filename)
     if not safe:
         flash('Invalid database selection.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     path = os.path.join(app.get_db_folder(), safe)
     if not os.path.isfile(path):
         flash('Database not found.', 'error')
-        return redirect(url_for('index'))
+        return redirect(url_for('core.index'))
     app.close_connection(None)
     temp_path = os.path.join(app.get_db_folder(), app.TEMP_DB_NAME)
     if app.app.config.get('DATABASE') == temp_path and os.path.exists(temp_path):
@@ -114,4 +114,4 @@ def load_saved_db():
         flash('Database loaded.', 'success')
     except Exception as e:
         flash(f'Error loading database: {e}', 'error')
-    return redirect(url_for('index'))
+    return redirect(url_for('core.index'))
