+++ retrorecon/routes/notes.py
@@ -1,21 +1,21 @@
-import app
 from flask import Blueprint, request, jsonify
+from ..services import db_service, tags as tags_service, notes as notes_service
 
 bp = Blueprint('notes', __name__)
 
 @bp.route('/saved_tags', methods=['GET', 'POST'])
 def saved_tags_route():
     if request.method == 'GET':
-        return jsonify({'tags': app.load_saved_tags()})
+        return jsonify({'tags': tags_service.load_saved_tags()})
     tag = request.form.get('tag', '').strip()
     if not tag:
         return ('', 400)
     if not tag.startswith('#'):
         tag = '#' + tag
-    tags = app.load_saved_tags()
+    tags = tags_service.load_saved_tags()
     if tag not in tags:
         tags.append(tag)
-        app.save_saved_tags(tags)
+        tags_service.save_saved_tags(tags)
     return ('', 204)
 
 @bp.route('/delete_saved_tag', methods=['POST'])
@@ -25,17 +25,17 @@ def delete_saved_tag():
         return ('', 400)
     if not tag.startswith('#'):
         tag = '#' + tag
-    tags = app.load_saved_tags()
+    tags = tags_service.load_saved_tags()
     if tag in tags:
         tags.remove(tag)
-        app.save_saved_tags(tags)
+        tags_service.save_saved_tags(tags)
     return ('', 204)
 
 @bp.route('/notes/<int:url_id>', methods=['GET'])
 def notes_get(url_id: int):
-    if not app._db_loaded():
+    if not db_service.db_loaded():
         return jsonify([])
-    rows = app.get_notes(url_id)
+    rows = notes_service.get_notes(url_id)
     return jsonify([
         {
             'id': r['id'],
@@ -49,7 +49,7 @@ def notes_get(url_id: int):
 
 @bp.route('/notes', methods=['POST'])
 def notes_post():
-    if not app._db_loaded():
+    if not db_service.db_loaded():
         return ('', 400)
     url_id = request.form.get('url_id', type=int)
     content = request.form.get('content', '').strip()
@@ -57,9 +57,9 @@ def notes_post():
         return ('', 400)
     note_id = request.form.get('note_id', type=int)
     if note_id:
-        app.update_note(note_id, content)
+        notes_service.update_note(note_id, content)
     else:
-        app.add_note(url_id, content)
+        notes_service.add_note(url_id, content)
     return ('', 204)
 
 @bp.route('/delete_note', methods=['POST'])
@@ -68,16 +68,16 @@ def delete_note_route():
     url_id = request.form.get('url_id', type=int)
     delete_all = request.form.get('all', '0') == '1'
     if note_id:
-        app.delete_note_entry(note_id)
+        notes_service.delete_note_entry(note_id)
     elif url_id and delete_all:
-        app.delete_all_notes(url_id)
+        notes_service.delete_all_notes(url_id)
     else:
         return ('', 400)
     return ('', 204)
 
 @bp.route('/export_notes', methods=['GET'])
 def export_notes():
-    if not app._db_loaded():
+    if not db_service.db_loaded():
         return jsonify([])
-    data = app.export_notes_data()
+    data = notes_service.export_notes_data()
     return jsonify(data)
